requirements.txt
python-docx~=1.2.0
Flask~=3.1.2
flask-cors~=6.0.1
Werkzeug~=3.1.3
gTTS~=2.5.4